+++ lib/ooxml_parser/docx_parser/document_structure/default_style_helper.rb
@@ -53,6 +53,8 @@ module OoxmlParser
       DocxParagraph.new.parse_paragraph_style(@styles.default_style(:paragraph)
                                                      .paragraph_properties_node,
                                               DocumentStructure.default_run_style)
+      return unless @styles&.default_style(:paragraph)&.run_properties_node
+
       DocumentStructure.default_run_style
                        .parse_properties(@styles.default_style(:paragraph)
                                                 .run_properties_node)